// src/types/enums.ts
export enum SendMoneyPageState {
  FORM = 'FORM',
  RECIPIENT_DETAILS = 'RECIPIENT_DETAILS',
  SENDER_DETAILS = 'SENDER_DETAILS',
  SUMMARY = 'SUMMARY',
}

export enum SendAirtimePageState {
  RECIPIENT_DETAILS = 'RECIPIENT_DETAILS',
  AMOUNT_FORM = 'AMOUNT_FORM',
  SENDER_DETAILS = 'SENDER_DETAILS',
  SUMMARY = 'SUMMARY',
}